heavy refactoring, added boss health graph to boss summary, dps graph and boss health are functional, missing mechanics and boons
